+++ Changelog.md
@@ -1,4 +1,6 @@
 Change log:
+* 2023-10-17 - U37 - 2 new + 1 newly available
+  * Goat of Forest-sprirt, Parrot, Fennec Fox
 * 2023-08-31 - U37 - 1 new + 1 future
   * new: Butterflies
   * furture: Fennec Fox